Conflicts:
	front/AVR/Hardware/SD/SPI_routines.c
	front/AVR/IO/uart.c
	front/AVR/Menu/menu.c
	mainboard/firmware/DrumSynth_FPU/Debug/makefile
	mainboard/firmware/DrumSynth_FPU/Debug/sources.mk
	mainboard/firmware/DrumSynth_FPU/src/DSPAudio/Oscillator.h
	mainboard/firmware/DrumSynth_FPU/src/MIDI/MidiMessages.h
	mainboard/firmware/DrumSynth_FPU/src/MIDI/frontPanelParser.c
	mainboard/firmware/DrumSynth_FPU/src/SampleRom/SampleMemory.c
	mainboard/firmware/DrumSynth_FPU/src/config.h
	mainboard/firmware/DrumSynth_FPU/src/main.c
	mainboard/firmware/UartBootloader_FPU/stm32_flash.ld

+++ front/AVR/Preset/PresetManager.h
@@ -36,7 +36,7 @@ uint8_t preset_loadPattern(uint8_t presetNr);
 char* preset_getPatternName(uint8_t presetNr);
 /** morph pattern linear to üpattern buffer 2*/
 void preset_morph(uint8_t morph);
-uint8_t preset_getMorphValue(uint8_t index, uint8_t morph);
+uint8_t preset_getMorphValue(uint16_t index, uint8_t morph);
 
 
 #endif /* PRESETMANAGER_H_ */
